Align the usage of RSpace scheduler for all public methods
