+++ node/src/test/scala/coop/rchain/node/configuration/commandline/Ed25519PubKeyConverterSpec.scala
@@ -13,9 +13,9 @@ class Ed25519PubKeyConverterSpec extends PropSpec with GeneratorDrivenPropertyCh
     val gen = Gen.listOfN(Ed25519.publicKeyLength, Arbitrary.arbByte.arbitrary).map(_.toArray)
 
     forAll(gen) { bytes: Array[Byte] =>
-      val asHex           = Base16.encode(bytes)
+      val asHex = Base16.encode(bytes)
 
-      def getResult(s : String) =
+      def getResult(s: String) =
         Ed25519PubKeyConverter.parse(List(("--user-id", List(s)))).right.get.get.bytes
 
       getResult(asHex) should be(bytes)
@@ -40,7 +40,9 @@ class Ed25519PubKeyConverterSpec extends PropSpec with GeneratorDrivenPropertyCh
     forAll { s: String =>
       val args = List(("--user-id", List(s)))
 
-      val isBadEncoding = s.isEmpty || s.exists(c => !(c >= '0' && c <= '9') || (c >= 'a' && c <= 'f'))
+      val isBadEncoding = s.isEmpty || s.exists(
+        c => !(c >= '0' && c <= '9') || (c >= 'a' && c <= 'f')
+      )
 
       Ed25519PubKeyConverter.parse(args).isLeft should be(isBadEncoding)
     }
